+++ portable-hash/src/ord_impls/ord_std.rs
@@ -1,7 +1,4 @@
-use std::{
-    net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6},
-    time::{Instant, SystemTime},
-};
+use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6};
 use crate::PortableOrd;
 
 // TODO(stabilisation): do we trust enum ordering is stable for IpAddr?
@@ -32,16 +29,6 @@ impl PortableOrd for SocketAddrV4 {
 }
 
 impl PortableOrd for SocketAddrV6 {
-    const CAN_USE_UNSTABLE_SORT: bool = true;
-    const I_KNOW_WHAT_I_AM_DOING: () = ();
-}
-
-impl PortableOrd for Instant {
-    const CAN_USE_UNSTABLE_SORT: bool = true;
-    const I_KNOW_WHAT_I_AM_DOING: () = ();
-}
-
-impl PortableOrd for SystemTime {
     const CAN_USE_UNSTABLE_SORT: bool = true;
     const I_KNOW_WHAT_I_AM_DOING: () = ();
 }